refactor: split rdpdr_backend.rs into smaller modules
Split the 1268-line rdpdr_backend.rs into a rdpdr/ directory with
focused submodules:
- mod.rs: Core types, MultiDriveBackend struct, RdpdrBackend impl
- file_ops.rs: File create, read, write, close operations
- query_ops.rs: Query information, volume info, directory listing
- set_ops.rs: Set information (rename, delete disposition, etc.)
- helpers.rs: Platform-specific helpers (time conversion, disk space)

Also includes fix for file handle metadata: prefer File::metadata()
over fs::metadata(path) so operations work even if file is deleted
locally while Windows still has a handle open.
